[s2296] Fix crash when using DB script command 34 (TERMINATE_CONDITION).
Using DB script command 34 (TERMINATE_CONDITION) was triggering an array out of
bounds access when debug logging was on, causing the server to crash. The array
is accessed with ConditionSource as index, but there was no corresponding string
for CONDITION_FROM_RESERVED_1. As a result, the array was missing one string and
accessing the array with the highest possible index, CONDITION_FROM_DBSCRIPTS,
accessed the array out of bounds.

Many thanks to @Schmoozerd for debugging and suggesting a fix!

Fixes cmangos/issues#591.